+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- add method verifyEmail
 
 ## [1.34.1] - 2025-01-07
 - verifyPasswordless should not make redirection with useWebMessage by using POST method endpoint.